towards schema based prompt file validation
