+++ src/vs/workbench/contrib/chat/common/promptSyntax/languageProviders/PromptHeaderDefinitionProvider.ts
@@ -9,8 +9,8 @@ import { Range } from '../../../../../../editor/common/core/range.js';
 import { Definition, DefinitionProvider } from '../../../../../../editor/common/languages.js';
 import { ITextModel } from '../../../../../../editor/common/model.js';
 import { IChatModeService } from '../../chatModes.js';
-import { getPromptsTypeForLanguageId } from '../promptTypes.js';
 import { PromptHeaderAttributes } from '../promptFileParser.js';
+import { getPromptsTypeForLanguageId } from '../promptTypes.js';
 import { IPromptsService } from '../service/promptsService.js';
 
 export class PromptHeaderDefinitionProvider implements DefinitionProvider {
