+++ src/vs/workbench/contrib/chat/common/promptSyntax/languageProviders/promptValidator.ts
@@ -16,18 +16,18 @@ import { ChatModeKind } from '../../constants.js';
 import { ILanguageModelChatMetadata, ILanguageModelsService } from '../../languageModels.js';
 import { ILanguageModelToolsService, SpecedToolAliases } from '../../tools/languageModelToolsService.js';
 import { getPromptsTypeForLanguageId, PromptsType } from '../promptTypes.js';
-import { GithubPromptHeaderAttributes, ISequenceValue, IHeaderAttribute, IScalarValue, parseCommaSeparatedList, ParsedPromptFile, PromptHeader, PromptHeaderAttributes, IValue } from '../promptFileParser.js';
+import { ISequenceValue, IHeaderAttribute, IScalarValue, parseCommaSeparatedList, ParsedPromptFile, PromptHeader, IValue, PromptHeaderAttributes } from '../promptFileParser.js';
 import { Disposable, DisposableStore, toDisposable } from '../../../../../../base/common/lifecycle.js';
 import { Delayer } from '../../../../../../base/common/async.js';
 import { ResourceMap } from '../../../../../../base/common/map.js';
 import { IFileService } from '../../../../../../platform/files/common/files.js';
 import { IPromptsService, Target } from '../service/promptsService.js';
 import { ILabelService } from '../../../../../../platform/label/common/label.js';
-import { AGENTS_SOURCE_FOLDER, CLAUDE_AGENTS_SOURCE_FOLDER, isInClaudeRulesFolder, LEGACY_MODE_FILE_EXTENSION } from '../config/promptFileLocations.js';
+import { AGENTS_SOURCE_FOLDER, LEGACY_MODE_FILE_EXTENSION } from '../config/promptFileLocations.js';
 import { Lazy } from '../../../../../../base/common/lazy.js';
 import { CancellationToken } from '../../../../../../base/common/cancellation.js';
-import { dirname } from '../../../../../../base/common/resources.js';
-import { URI } from '../../../../../../base/common/uri.js';
+import { claudeAgentAttributes, getTarget, getValidAttributeNames, GithubPromptHeaderAttributes, isVSCodeOrDefaultTarget } from './promptFileAttributes.js';
+
 
 export const MARKERS_OWNER_ID = 'prompts-diagnostics-provider';
 
@@ -755,291 +755,6 @@ function isTrueOrFalse(value: IValue): boolean {
 		return (value.value === 'true' || value.value === 'false') && value.format === 'none';
 	}
 	return false;
-}
-
-const allAttributeNames: Record<PromptsType, string[]> = {
-	[PromptsType.prompt]: [PromptHeaderAttributes.name, PromptHeaderAttributes.description, PromptHeaderAttributes.model, PromptHeaderAttributes.tools, PromptHeaderAttributes.mode, PromptHeaderAttributes.agent, PromptHeaderAttributes.argumentHint],
-	[PromptsType.instructions]: [PromptHeaderAttributes.name, PromptHeaderAttributes.description, PromptHeaderAttributes.applyTo, PromptHeaderAttributes.excludeAgent],
-	[PromptsType.agent]: [PromptHeaderAttributes.name, PromptHeaderAttributes.description, PromptHeaderAttributes.model, PromptHeaderAttributes.tools, PromptHeaderAttributes.advancedOptions, PromptHeaderAttributes.handOffs, PromptHeaderAttributes.argumentHint, PromptHeaderAttributes.target, PromptHeaderAttributes.infer, PromptHeaderAttributes.agents, PromptHeaderAttributes.userInvocable, PromptHeaderAttributes.userInvokable, PromptHeaderAttributes.disableModelInvocation, GithubPromptHeaderAttributes.github],
-	[PromptsType.skill]: [PromptHeaderAttributes.name, PromptHeaderAttributes.description, PromptHeaderAttributes.license, PromptHeaderAttributes.compatibility, PromptHeaderAttributes.metadata, PromptHeaderAttributes.argumentHint, PromptHeaderAttributes.userInvocable, PromptHeaderAttributes.userInvokable, PromptHeaderAttributes.disableModelInvocation],
-	[PromptsType.hook]: [], // hooks are JSON files, not markdown with YAML frontmatter
-};
-const githubCopilotAgentAttributeNames = [PromptHeaderAttributes.name, PromptHeaderAttributes.description, PromptHeaderAttributes.tools, PromptHeaderAttributes.target, GithubPromptHeaderAttributes.mcpServers, GithubPromptHeaderAttributes.github, PromptHeaderAttributes.infer];
-const recommendedAttributeNames: Record<PromptsType, string[]> = {
-	[PromptsType.prompt]: allAttributeNames[PromptsType.prompt].filter(name => !isNonRecommendedAttribute(name)),
-	[PromptsType.instructions]: allAttributeNames[PromptsType.instructions].filter(name => !isNonRecommendedAttribute(name)),
-	[PromptsType.agent]: allAttributeNames[PromptsType.agent].filter(name => !isNonRecommendedAttribute(name)),
-	[PromptsType.skill]: allAttributeNames[PromptsType.skill].filter(name => !isNonRecommendedAttribute(name)),
-	[PromptsType.hook]: [], // hooks are JSON files, not markdown with YAML frontmatter
-};
-
-export function getValidAttributeNames(promptType: PromptsType, includeNonRecommended: boolean, target: Target): string[] {
-	if (target === Target.Claude) {
-		if (promptType === PromptsType.instructions) {
-			return Object.keys(claudeRulesAttributes);
-		}
-		return Object.keys(claudeAgentAttributes);
-	} else if (target === Target.GitHubCopilot) {
-		if (promptType === PromptsType.agent) {
-			return githubCopilotAgentAttributeNames;
-		}
-	}
-	return includeNonRecommended ? allAttributeNames[promptType] : recommendedAttributeNames[promptType];
-}
-
-export function isNonRecommendedAttribute(attributeName: string): boolean {
-	return attributeName === PromptHeaderAttributes.advancedOptions || attributeName === PromptHeaderAttributes.excludeAgent || attributeName === PromptHeaderAttributes.mode || attributeName === PromptHeaderAttributes.infer || attributeName === PromptHeaderAttributes.userInvokable;
-}
-
-export function getAttributeDescription(attributeName: string, promptType: PromptsType, target: Target): string | undefined {
-	if (target === Target.Claude) {
-		if (promptType === PromptsType.agent) {
-			return claudeAgentAttributes[attributeName]?.description;
-		}
-		if (promptType === PromptsType.instructions) {
-			return claudeRulesAttributes[attributeName]?.description;
-		}
-	}
-	switch (promptType) {
-		case PromptsType.instructions:
-			switch (attributeName) {
-				case PromptHeaderAttributes.name:
-					return localize('promptHeader.instructions.name', 'The name of the instruction file as shown in the UI. If not set, the name is derived from the file name.');
-				case PromptHeaderAttributes.description:
-					return localize('promptHeader.instructions.description', 'The description of the instruction file. It can be used to provide additional context or information about the instructions and is passed to the language model as part of the prompt.');
-				case PromptHeaderAttributes.applyTo:
-					return localize('promptHeader.instructions.applyToRange', 'One or more glob pattern (separated by comma) that describe for which files the instructions apply to. Based on these patterns, the file is automatically included in the prompt, when the context contains a file that matches one or more of these patterns. Use `**` when you want this file to always be added.\nExample: `**/*.ts`, `**/*.js`, `client/**`');
-			}
-			break;
-		case PromptsType.skill:
-			switch (attributeName) {
-				case PromptHeaderAttributes.name:
-					return localize('promptHeader.skill.name', 'The name of the skill.');
-				case PromptHeaderAttributes.description:
-					return localize('promptHeader.skill.description', 'The description of the skill. The description is added to every request and will be used by the agent to decide when to load the skill.');
-				case PromptHeaderAttributes.argumentHint:
-					return localize('promptHeader.skill.argumentHint', 'Hint shown during autocomplete to indicate expected arguments. Example: [issue-number] or [filename] [format]');
-				case PromptHeaderAttributes.userInvocable:
-					return localize('promptHeader.skill.userInvocable', 'Set to false to hide from the / menu. Use for background knowledge users should not invoke directly. Default: true.');
-				case PromptHeaderAttributes.disableModelInvocation:
-					return localize('promptHeader.skill.disableModelInvocation', 'Set to true to prevent the agent from automatically loading this skill. Use for workflows you want to trigger manually with /name. Default: false.');
-			}
-			break;
-		case PromptsType.agent:
-			switch (attributeName) {
-				case PromptHeaderAttributes.name:
-					return localize('promptHeader.agent.name', 'The name of the agent as shown in the UI.');
-				case PromptHeaderAttributes.description:
-					return localize('promptHeader.agent.description', 'The description of the custom agent, what it does and when to use it.');
-				case PromptHeaderAttributes.argumentHint:
-					return localize('promptHeader.agent.argumentHint', 'The argument-hint describes what inputs the custom agent expects or supports.');
-				case PromptHeaderAttributes.model:
-					return localize('promptHeader.agent.model', 'Specify the model that runs this custom agent. Can also be a list of models. The first available model will be used.');
-				case PromptHeaderAttributes.tools:
-					return localize('promptHeader.agent.tools', 'The set of tools that the custom agent has access to.');
-				case PromptHeaderAttributes.handOffs:
-					return localize('promptHeader.agent.handoffs', 'Possible handoff actions when the agent has completed its task.');
-				case PromptHeaderAttributes.target:
-					return localize('promptHeader.agent.target', 'The target to which the header attributes like tools apply to. Possible values are `github-copilot` and `vscode`.');
-				case PromptHeaderAttributes.infer:
-					return localize('promptHeader.agent.infer', 'Controls visibility of the agent.');
-				case PromptHeaderAttributes.agents:
-					return localize('promptHeader.agent.agents', 'One or more agents that this agent can use as subagents. Use \'*\' to specify all available agents.');
-				case PromptHeaderAttributes.userInvocable:
-					return localize('promptHeader.agent.userInvocable', 'Whether the agent can be selected and invoked by users in the UI.');
-				case PromptHeaderAttributes.disableModelInvocation:
-					return localize('promptHeader.agent.disableModelInvocation', 'If true, prevents the agent from being invoked as a subagent.');
-				case GithubPromptHeaderAttributes.github:
-					return localize('promptHeader.agent.github', 'GitHub-specific configuration for the agent, such as token permissions.');
-			}
-			break;
-		case PromptsType.prompt:
-			switch (attributeName) {
-				case PromptHeaderAttributes.name:
-					return localize('promptHeader.prompt.name', 'The name of the prompt. This is also the name of the slash command that will run this prompt.');
-				case PromptHeaderAttributes.description:
-					return localize('promptHeader.prompt.description', 'The description of the reusable prompt, what it does and when to use it.');
-				case PromptHeaderAttributes.argumentHint:
-					return localize('promptHeader.prompt.argumentHint', 'The argument-hint describes what inputs the prompt expects or supports.');
-				case PromptHeaderAttributes.model:
-					return localize('promptHeader.prompt.model', 'The model to use in this prompt. Can also be a list of models. The first available model will be used.');
-				case PromptHeaderAttributes.tools:
-					return localize('promptHeader.prompt.tools', 'The tools to use in this prompt.');
-				case PromptHeaderAttributes.agent:
-				case PromptHeaderAttributes.mode:
-					return localize('promptHeader.prompt.agent.description', 'The agent to use when running this prompt.');
-			}
-			break;
-	}
-	return undefined;
-}
-
-// The list of tools known to be used by GitHub Copilot custom agents
-export const knownGithubCopilotTools = [
-	{ name: SpecedToolAliases.execute, description: localize('githubCopilot.execute', 'Execute commands') },
-	{ name: SpecedToolAliases.read, description: localize('githubCopilot.read', 'Read files') },
-	{ name: SpecedToolAliases.edit, description: localize('githubCopilot.edit', 'Edit files') },
-	{ name: SpecedToolAliases.search, description: localize('githubCopilot.search', 'Search files') },
-	{ name: SpecedToolAliases.agent, description: localize('githubCopilot.agent', 'Use subagents') },
-];
-
-export interface IValueEntry {
-	readonly name: string;
-	readonly description?: string;
-}
-
-export const knownClaudeTools = [
-	{ name: 'Bash', description: localize('claude.bash', 'Execute shell commands'), toolEquivalent: [SpecedToolAliases.execute] },
-	{ name: 'Edit', description: localize('claude.edit', 'Make targeted file edits'), toolEquivalent: ['edit/editNotebook', 'edit/editFiles'] },
-	{ name: 'Glob', description: localize('claude.glob', 'Find files by pattern'), toolEquivalent: ['search/fileSearch'] },
-	{ name: 'Grep', description: localize('claude.grep', 'Search file contents with regex'), toolEquivalent: ['search/textSearch'] },
-	{ name: 'Read', description: localize('claude.read', 'Read file contents'), toolEquivalent: ['read/readFile', 'read/getNotebookSummary'] },
-	{ name: 'Write', description: localize('claude.write', 'Create/overwrite files'), toolEquivalent: ['edit/createDirectory', 'edit/createFile', 'edit/createJupyterNotebook'] },
-	{ name: 'WebFetch', description: localize('claude.webFetch', 'Fetch URL content'), toolEquivalent: [SpecedToolAliases.web] },
-	{ name: 'WebSearch', description: localize('claude.webSearch', 'Perform web searches'), toolEquivalent: [SpecedToolAliases.web] },
-	{ name: 'Task', description: localize('claude.task', 'Run subagents for complex tasks'), toolEquivalent: [SpecedToolAliases.agent] },
-	{ name: 'Skill', description: localize('claude.skill', 'Execute skills'), toolEquivalent: [] },
-	{ name: 'LSP', description: localize('claude.lsp', 'Code intelligence (requires plugin)'), toolEquivalent: [] },
-	{ name: 'NotebookEdit', description: localize('claude.notebookEdit', 'Modify Jupyter notebooks'), toolEquivalent: ['edit/editNotebook'] },
-	{ name: 'AskUserQuestion', description: localize('claude.askUserQuestion', 'Ask multiple-choice questions'), toolEquivalent: ['vscode/askQuestions'] },
-	{ name: 'MCPSearch', description: localize('claude.mcpSearch', 'Searches for MCP tools when tool search is enabled'), toolEquivalent: [] }
-];
-
-export const knownClaudeModels = [
-	{ name: 'sonnet', description: localize('claude.sonnet', 'Latest Claude Sonnet'), modelEquivalent: 'Claude Sonnet 4.5 (copilot)' },
-	{ name: 'opus', description: localize('claude.opus', 'Latest Claude Opus'), modelEquivalent: 'Claude Opus 4.6 (copilot)' },
-	{ name: 'haiku', description: localize('claude.haiku', 'Latest Claude Haiku, fast for simple tasks'), modelEquivalent: 'Claude Haiku 4.5 (copilot)' },
-	{ name: 'inherit', description: localize('claude.inherit', 'Inherit model from parent agent or prompt'), modelEquivalent: undefined },
-];
-
-export function mapClaudeModels(claudeModelNames: readonly string[]): readonly string[] {
-	const result = [];
-	for (const name of claudeModelNames) {
-		const claudeModel = knownClaudeModels.find(model => model.name === name);
-		if (claudeModel && claudeModel.modelEquivalent) {
-			result.push(claudeModel.modelEquivalent);
-		}
-	}
-	return result;
-}
-
-/**
- * Maps Claude tool names to their VS Code tool equivalents.
- */
-export function mapClaudeTools(claudeToolNames: readonly string[]): string[] {
-	const result: string[] = [];
-	for (const name of claudeToolNames) {
-		const claudeTool = knownClaudeTools.find(tool => tool.name === name);
-		if (claudeTool) {
-			result.push(...claudeTool.toolEquivalent);
-		}
-	}
-	return result;
-}
-
-export const claudeAgentAttributes: Record<string, { type: string; description: string; defaults?: string[]; items?: IValueEntry[]; enums?: IValueEntry[] }> = {
-	'name': {
-		type: 'scalar',
-		description: localize('attribute.name', "Unique identifier using lowercase letters and hyphens (required)"),
-	},
-	'description': {
-		type: 'scalar',
-		description: localize('attribute.description', "When to delegate to this subagent (required)"),
-	},
-	'tools': {
-		type: 'sequence',
-		description: localize('attribute.tools', "Array of tools the subagent can use. Inherits all tools if omitted"),
-		defaults: ['Read, Edit, Bash'],
-		items: knownClaudeTools
-	},
-	'disallowedTools': {
-		type: 'sequence',
-		description: localize('attribute.disallowedTools', "Tools to deny, removed from inherited or specified list"),
-		defaults: ['Write, Edit, Bash'],
-		items: knownClaudeTools
-	},
-	'model': {
-		type: 'scalar',
-		description: localize('attribute.model', "Model to use: sonnet, opus, haiku, or inherit. Defaults to inherit."),
-		defaults: ['sonnet', 'opus', 'haiku', 'inherit'],
-		enums: knownClaudeModels
-	},
-	'permissionMode': {
-		type: 'scalar',
-		description: localize('attribute.permissionMode', "Permission mode: default, acceptEdits, dontAsk, bypassPermissions, or plan."),
-		defaults: ['default', 'acceptEdits', 'dontAsk', 'bypassPermissions', 'plan'],
-		enums: [
-			{ name: 'default', description: localize('claude.permissionMode.default', 'Standard behavior: prompts for permission on first use of each tool.') },
-			{ name: 'acceptEdits', description: localize('claude.permissionMode.acceptEdits', 'Automatically accepts file edit permissions for the session.') },
-			{ name: 'plan', description: localize('claude.permissionMode.plan', 'Plan Mode: Claude can analyze but not modify files or execute commands.') },
-			{ name: 'delegate', description: localize('claude.permissionMode.delegate', 'Coordination-only mode for agent team leads. Only available when an agent team is active.') },
-			{ name: 'dontAsk', description: localize('claude.permissionMode.dontAsk', 'Auto-denies tools unless pre-approved via /permissions or permissions.allow rules.') },
-			{ name: 'bypassPermissions', description: localize('claude.permissionMode.bypassPermissions', 'Skips all permission prompts (requires safe environment like containers).') }
-		]
-	},
-	'skills': {
-		type: 'sequence',
-		description: localize('attribute.skills', "Skills to load into the subagent's context at startup."),
-	},
-	'mcpServers': {
-		type: 'sequence',
-		description: localize('attribute.mcpServers', "MCP servers available to this subagent."),
-	},
-	'hooks': {
-		type: 'object',
-		description: localize('attribute.hooks', "Lifecycle hooks scoped to this subagent."),
-	},
-	'memory': {
-		type: 'scalar',
-		description: localize('attribute.memory', "Persistent memory scope: user, project, or local. Enables cross-session learning."),
-		defaults: ['user', 'project', 'local'],
-		enums: [
-			{ name: 'user', description: localize('claude.memory.user', "Remember learnings across all projects.") },
-			{ name: 'project', description: localize('claude.memory.project', "The subagent's knowledge is project-specific and shareable via version control.") },
-			{ name: 'local', description: localize('claude.memory.local', "The subagent's knowledge is project-specific but should not be checked into version control.") }
-		]
-	}
-};
-
-/**
- * Attributes supported in Claude rules files (`.claude/rules/*.md`).
- * Claude rules use `paths` instead of `applyTo` for glob patterns.
- */
-export const claudeRulesAttributes: Record<string, { type: string; description: string; defaults?: string[]; items?: IValueEntry[]; enums?: IValueEntry[] }> = {
-	'description': {
-		type: 'scalar',
-		description: localize('attribute.rules.description', "A description of what this rule covers, used to provide context about when it applies."),
-	},
-	'paths': {
-		type: 'sequence',
-		description: localize('attribute.rules.paths', "Array of glob patterns that describe for which files the rule applies. Based on these patterns, the file is automatically included in the prompt when the context contains a file that matches.\nExample: `['src/**/*.ts', 'test/**']`"),
-	},
-};
-
-export function isVSCodeOrDefaultTarget(target: Target): boolean {
-	return target === Target.VSCode || target === Target.Undefined;
-}
-
-export function getTarget(promptType: PromptsType, header: PromptHeader | URI): Target {
-	const uri = header instanceof URI ? header : header.uri;
-	if (promptType === PromptsType.agent) {
-		const parentDir = dirname(uri);
-		if (parentDir.path.endsWith(`/${CLAUDE_AGENTS_SOURCE_FOLDER}`)) {
-			return Target.Claude;
-		}
-		if (!(header instanceof URI)) {
-			const target = header.target;
-			if (target === Target.GitHubCopilot || target === Target.VSCode) {
-				return target;
-			}
-		}
-		return Target.Undefined;
-	} else if (promptType === PromptsType.instructions) {
-		if (isInClaudeRulesFolder(uri)) {
-			return Target.Claude;
-		}
-	}
-	return Target.Undefined;
 }
 
 function toMarker(message: string, range: Range, severity = MarkerSeverity.Error): IMarkerData {
